==> 90.results.Rmd <==
---
title: "90_results"
author: "Guillaume Papuga"
date: "28/04/2021"
output: html_document
---
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
```

 
# Figure X : temporal dissimilarity
```{r cars}
t = imager::load.image(here::here ("outputs", "figures", "temp.dissi.jpg"))
plot(t, axes = FALSE)

```
Analyse the plot here.
